feat(inspector): rewrite the 3D inspector
Summary: The 3D Inspector's logic was not well organized and was deeply intertwined with the device transform display logic and the resizing mask logic. This diff is a rewrite of the 3D Inspector, which separates it from the other components while providing methods for other parts of the emulator to access information and control the emulated device and scene rendering.

Reviewed By: cabanier

Differential Revision:
D46954731

Privacy Context Container: L1164747

fbshipit-source-id: a6460f4a86136b44a03be95e48ee43c9cf395fc4