fix(three-version-leak): Prevent THREE version from leaking into content
Summary: Resolved the issue where the THREE version used in the Immersive Web Emulator was leaking into the content by moving to a custom build of THREE that does not define properties on window. This was causing multiple instances of Three.js to be imported, as indicated by the warning in A-Frame. The fix ensures that by the time A-Frame loads, only the appropriate THREE revision is defined, preventing any version conflicts.

Reviewed By: cabanier

Differential Revision: D51960562

fbshipit-source-id: d483569d1e676e8b338464be326269b19153e8fb