feat(user-objects): Add visibility toggle for user objects
Summary:
Resolves the issue outlined in https://github.com/meta-quest/immersive-web-emulator/issues/50.
This update introduces a visibility toggle for user-defined objects, including planes and meshes. When a user-defined object is set to hidden, it will display as transparent in the emulator UI, and a '[hidden]' tag will precede its semantic label. Concurrently, the corresponding XRPlane or XRMesh will be hidden from the WebXR application until the visibility is toggled back on.

Reviewed By: cabanier

Differential Revision: D51968463

fbshipit-source-id: abe1a0075ec600a393208d96c94c97d46a50b6ab