Added controller hide button to hide controller from the emulator UI and polyfill injection.
Summary:
Added 2 buttons to hide the controller panel from the emulator UI as well as blocking the controller signal from injecting to the polyfill. The controller mesh will also be hidden from the emulator UI.

V2: Addressed comments and not hiding the controller panel.

V3: Added a new class in the style sheet to group the controller icon and name.

Reviewed By: felixtrz

Differential Revision: D54088960

fbshipit-source-id: ac31019aea8d9158ba7b8c564f86ed833eec871f